fix: distribute rewards with wallet connect
