+++ package.json
@@ -76,7 +76,8 @@
     "redux-devtools-extension": "^2.13.8",
     "styled-components": "^5.2.1",
     "swr": "^0.4.2",
-    "web3": "^1.5.2"
+    "web3": "^1.5.2",
+    "web3-eth-contract": "^1.6.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
