+++ src/modules/hooks/index.ts
@@ -15,6 +15,7 @@ export { useGetSbBtcBal } from './useGetSbBtcBal';
 export { useGetLatestPrice } from './useGetLatestPrice';
 export { useGetEarningHistorical } from './useGetEarningHistorical';
 export { useGetSwapRewards } from './useGetSwapRewards';
+export { useDistributeRewards } from './useDistributeRewards';
 
 export interface ICountdown {
   days: number;
